--- src/extensions/cli/cli.extension.ts
@@ -1,9 +1,0 @@
-import { ExtensionManifest } from '@teambit/harmony';
-import { CLIProvider } from './cli.provider';
-import { PaperExt } from '../paper';
-
-export const BitCliExt: ExtensionManifest = {
-  name: 'cli',
-  dependencies: [PaperExt],
-  provider: CLIProvider
-};
