--- src/extensions/cli/index.ts
@@ -1,3 +1,0 @@
-export { BitCliExt } from './cli.extension';
-export { BitCli } from './cli';
-export { Command, CLIArgs } from '../paper';
