+++ src/extensions/paper/legacy-command.tsx
@@ -1,11 +1,11 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { Color, AppContext } from 'ink';
 import React from 'react';
-import { Command, PaperOptions, GenericObject } from '../paper';
+import { Command, PaperOptions, GenericObject } from '.';
 import LegacyInterface from '../../cli/command';
 import allHelp from '../../cli/templates/all-help';
-import { getID } from '../paper';
-import { Paper } from '../paper';
+import { getID } from '.';
+import { PaperExtension } from '.';
 
 export class LegacyCommand implements Command {
   alias: string;
@@ -18,7 +18,7 @@ export class LegacyCommand implements Command {
   commands: Command[];
   private?: boolean;
   migration?: boolean;
-  constructor(private cmd: LegacyInterface, p: Paper) {
+  constructor(private cmd: LegacyInterface, p: PaperExtension) {
     this.name = cmd.name;
     this.description = cmd.description;
     this.options = cmd.opts;
@@ -66,7 +66,7 @@ export class LegacyCommand implements Command {
 }
 
 // TODO: remove all help and move information to commands
-export function findLegacyDetails(name: string, p: Paper) {
+export function findLegacyDetails(name: string, p: PaperExtension) {
   let group = '';
   let summery = '';
   for (let i = 0; i < allHelp.length; i += 1) {
